// libs/cart/driver/src/errors/codes.enum.ts
export enum DaffCartDriverErrorCodes {
  CART_NOT_FOUND = 'DAFF_CART_NOT_FOUND',
  PRODUCT_OUT_OF_STOCK = 'DAFF_PRODUCT_OUT_OF_STOCK',
  INVALID_COUPON_CODE = 'DAFF_INVALID_COUPON_CODE',
  INVALID_COUNTRY = 'DAFF_INVALID_COUNTRY',
  INVALID_EMAIL = 'DAFF_INVALID_EMAIL',
  INVALID_REGION = 'DAFF_INVALID_REGION',
  INVALID_API_RESPONSE = 'DAFF_INVALID_API_RESPONSE',
  EXPIRED_PAYMENT_METHOD = 'DAFF_EXPIRED_PAYMENT_METHOD',
  ITEM_EXCEEDS_MAX_QTY = 'ITEM_EXCEEDS_MAX_QTY',
  UNAUTHORIZED_FOR_CART = 'DAFF_UNAUTHORIZED_FOR_CART',
}